chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/html/imports/import-events-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE ERROR: line 19: Uncaught TypeError: Cannot read property 'documentElement' of undefined
Test that events can be fired on imported documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
